+++ Address/src/main/java/com/ondemandcarwash/repository/AddressRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.ondemandcarwash.model.Address;
 
-public interface AddressRepository extends MongoRepository<Address, Integer>{
+public interface AddressRepository extends MongoRepository<Address, Long>{
 
 }
